refactor(Graph): Move table data point and multiple y axes code to util
#1289
